pyproject.toml
[tool.poetry]
name = "doc2md"
version = "0.1.0"
description = "DOCX to Markdown converter pipeline"
authors = ["Example <example@example.com>"]
readme = "README.md"
packages = [{ include = "doc2md", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.3"
rich = "^13.7.0"
python-dotenv = "^1.0.1"
mammoth = "^1.6.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"
httpx = "^0.28.1"
jsonschema = "^4.25.1"
python-frontmatter = "^1.0.0"
python-slugify = "^8.0.4"
python-docx = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.6.3"
mypy = "^1.10.0"
pytest = "^8.2.1"
pytest-cov = "^6.2.1"

[tool.poetry.scripts]
doc2md = "doc2md.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
